backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
pillow==10.0.0
numpy<2
ultralytics>=8.0.0
